+++ samples/Store/Integration/CodecIntegration.fs
@@ -42,8 +42,8 @@ let codec = FsCodec.NewtonsoftJson.Codec.Create()
 
 [<AutoData(MaxTest=100)>]
 let ``Can roundtrip, rendering correctly`` (x: SimpleDu) =
-    let serialized = codec.Encode x
+    let serialized = codec.Encode(None,x)
     render x =! if serialized.Data = null then null else System.Text.Encoding.UTF8.GetString(serialized.Data)
-    let adapted = FsCodec.Core.TimelineEvent.Create(-1L, serialized.EventType, serialized.Data, serialized.Meta, serialized.Timestamp)
+    let adapted = FsCodec.Core.TimelineEvent.Create(-1L, serialized.EventType, serialized.Data)
     let deserialized = codec.TryDecode adapted |> Option.get
     deserialized =! x